claims_system/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
anthropic==0.34.2
openai==1.52.2
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.2
pymupdf==1.24.14
Pillow==10.4.0
pytest==8.3.3
